+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Set the default log retention to 30 days for full fidelity and downsampled data. ([#5065](https://github.com/getsentry/relay/pull/5065))
 - Improved PII Scrubbing for attributes / logs ([#5061](https://github.com/getsentry/relay/pull/5061)))
+- Introduces a project scope sampling rule type. ([#5077](https://github.com/getsentry/relay/pull/5077)))
 
 ## 25.8.0